move LoggerHolder to API there is a requirement (and some outstanding todo's) to log errors and warnings from the API, so move logger setup into API to facilitate this.
