+++ examples/traces/exporters/otlp_grpc_from_factory.php
@@ -6,7 +6,7 @@ use OpenTelemetry\SDK\Trace\TracerProviderFactory;
 
 require __DIR__ . '/../../../vendor/autoload.php';
 
-\OpenTelemetry\SDK\Common\Log\LoggerHolder::set(new \Monolog\Logger('grpc', [new \Monolog\Handler\StreamHandler('php://stderr')]));
+\OpenTelemetry\API\Common\Log\LoggerHolder::set(new \Monolog\Logger('grpc', [new \Monolog\Handler\StreamHandler('php://stderr')]));
 
 /**
  * Create an otlp+grpc tracer provider from TracerProviderFactory, using environment variables as input
