+++ examples/traces/exporters/otlp_http_from_factory.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 require __DIR__ . '/../../../vendor/autoload.php';
 
-\OpenTelemetry\SDK\Common\Log\LoggerHolder::set(new \Monolog\Logger('grpc', [new \Monolog\Handler\StreamHandler('php://stderr')]));
+\OpenTelemetry\API\Common\Log\LoggerHolder::set(new \Monolog\Logger('grpc', [new \Monolog\Handler\StreamHandler('php://stderr')]));
 
 /**
  * Create an otlp+http/protobuf tracer provider from TracerProviderFactory, using environment variables as input
