+++ src/Contrib/Jaeger/HttpSender.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace OpenTelemetry\Contrib\Jaeger;
 
 use Jaeger\Thrift\Process;
+use OpenTelemetry\API\Behavior\LogsMessagesTrait;
 use OpenTelemetry\Contrib\Jaeger\BatchAdapter\BatchAdapterFactory;
 use OpenTelemetry\Contrib\Jaeger\BatchAdapter\BatchAdapterFactoryInterface;
 use OpenTelemetry\Contrib\Jaeger\BatchAdapter\BatchAdapterInterface;
 use OpenTelemetry\Contrib\Jaeger\TagFactory\TagFactory;
-use OpenTelemetry\SDK\Behavior\LogsMessagesTrait;
 use OpenTelemetry\SDK\Resource\ResourceInfo;
 use OpenTelemetry\SDK\Resource\ResourceInfoFactory;
 use OpenTelemetry\SemConv\ResourceAttributes;
