+++ src/SDK/Common/Dev/Compatibility/_load.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 require_once __DIR__ . '/BC/ClockInterface.php';
 require_once __DIR__ . '/BC/AbstractClock.php';
 require_once __DIR__ . '/BC/SystemClock.php';
-require_once __DIR__ . '/BC/GlobalLoggerHolder.php';
 require_once __DIR__ . '/BC/InstrumentationLibraryInterface.php';
 require_once __DIR__ . '/BC/InstrumentationLibrary.php';
 require_once __DIR__ . '/BC/OtlpExporter.php';
+require_once __DIR__ . '/BC/LoggerHolder.php';
